+++ src/main/java/roomescape/controller/api/MemberApiController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Member;
-import roomescape.service.dto.response.MemberIdAndNameResponse;
+import roomescape.service.dto.response.member.MemberIdAndNameResponses;
 import roomescape.service.member.MemberService;
 
 @RestController
@@ -18,12 +18,8 @@ public class MemberApiController {
     }
 
     @GetMapping("/members")
-    public ResponseEntity<List<MemberIdAndNameResponse>> getMembers() {
+    public ResponseEntity<MemberIdAndNameResponses> getMembers() {
         List<Member> members = memberService.findMembers();
-        return ResponseEntity.ok(
-                members.stream()
-                        .map(member -> new MemberIdAndNameResponse(member.getId(), member.getName()))
-                        .toList()
-        );
+        return ResponseEntity.ok(MemberIdAndNameResponses.from(members));
     }
 }
