+++ src/main/java/roomescape/service/dto/response/member/MemberIdAndNameResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.response;
+package roomescape.service.dto.response.member;
 
 public record MemberIdAndNameResponse(Long id, String name) {
 }
