+++ src/main/resources/static/js/reservation-with-member.js
@@ -58,7 +58,7 @@ function fetchThemes() {
 function fetchMembers() {
     requestRead(MEMBER_API_ENDPOINT)
         .then(data => {
-            membersOptions.push(...data);
+            membersOptions.push(...data.members);
             populateSelect('member', membersOptions, 'name');
         })
         .catch(error => console.error('Error fetching member:', error));
