+++ src/test/java/roomescape/controller/api/AuthApiControllerTest.java
@@ -5,11 +5,10 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import roomescape.controller.BaseControllerTest;
 import roomescape.service.dto.request.LoginRequest;
-import roomescape.service.dto.response.MemberIdAndNameResponse;
+import roomescape.service.dto.response.member.MemberIdAndNameResponse;
 import roomescape.util.TokenGenerator;
 
 class AuthApiControllerTest extends BaseControllerTest {
